requirements.txt
fastapi
uvicorn[standard]
pandas
duckdb
python-dotenv
openai
plotly